the answer

+++ 0x01-variables_if_else_while/0-positive_or_negative.c
@@ -8,7 +8,7 @@
  */
 int main(void)
 {
-	int n = 0;
+	long int n = 0;
 
 	if (n > 0)
 	{
